Simplify message marshaling / unmarshaling
The duplicate "internal" types are gone. Instead, most types are
represented directly. Some types, those that use string values, have two
definitions, one private with string offsets and one public with actual
strings.
