--- prompting/apparmor/raw/raw.go
@@ -1,2 +1,0 @@
-// Package raw contains low-level kernel types and functions.
-package raw
