--- prompting/apparmor/raw/raw_test.go
@@ -1,13 +1,0 @@
-package raw_test
-
-import (
-	"testing"
-
-	. "gopkg.in/check.v1"
-)
-
-func Test(t *testing.T) { TestingT(t) }
-
-type miscSuite struct{}
-
-var _ = Suite(&miscSuite{})
